Added internal Timeout functions with persistance
Each state can have an individual timeout set. When that state is activated, it will automatically start a timer and will transition if the timer is allowed to expire.

The timer is persistent, so if the flow is restarted, the timer will still attempt to restart and expire at the same time as the original expiration.